refactoring of service folder and login/register fix
