+++ src/services/userService.js
@@ -1,5 +1,24 @@
-import User from "./../entities/User";
 import BASE_URL from './../shared/baseUrl';
+import Person from '../entities/Person';
+import User from "./../entities/User";
+
+const USERS_URL = '/users/?_embed=posts';
+
+const fetchUsers = () => {
+    return fetch(`${BASE_URL}${USERS_URL}`,
+        {
+            headers: {
+                "Content-Type": "application/ json",
+                "x-api-key": "B1tD3V"
+            }
+        }).then(response => response.json())
+        .then(response => response.map((user) => {
+            const { about, name, posts, avatarUrl, id } = user;
+            const lastPost = posts[posts.length - 1] ? posts[posts.length - 1].createdAt : "No posts to display";
+            return new Person(name.first, name.last, avatarUrl, (about || "").bio, lastPost, id)
+        })
+        )
+}
 
 
 const fetchUser = (id) => {
@@ -54,9 +73,7 @@ const editProfile = (id, fullName, imageUrl, bio) => {
         })
 }
 
-
-export {
-    fetchUser,
-    editProfile
-}
-
+export  {fetchUsers,
+        fetchUser,
+        editProfile
+        }
